requirements.txt
Django==3.1.3
Pillow==8.0.1
jinja2==3.0
gunicorn==20.1.0
contextvars
typing-extensions==4.1.1
aiohttp==3.8.1
click==8.0.4
wheel==0.37.1
multidict==4.5
django_nested_admin==4.0.2